Start of implementing Cameras, have controller and view showing up, model not in use yet. Minor edits to drt and vog view code.

+++ Devices/DRT/View/drt_view.py
@@ -43,7 +43,7 @@ class DRTView(AbstractView):
         super().__init__(name)
 
         self.subwindow_height = 450
-        self.tab_height = self.subwindow_height * 0.87
+        self.tab_height = int(self.subwindow_height * 0.87)
 
         # device settings display
         self.dev_sets_frame = EasyFrame()
